progress on employee jtable

+++ src/database/create/CreateEmployee.java
@@ -93,7 +93,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Feb"); // employee dob month
             pstmt.setString(10, "1984"); // employee dob year
             pstmt.setString(11, "ermac.ward@gmail.com");  //employee email
-            pstmt.setInt(12, 12000); // employee salary
+            pstmt.setDouble(12, 12000); // employee salary
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/ermac.png"))); // employee picture BLOB
             pstmt.setString(14, "ruthward"); // employee username
             pstmt.setString(15, "827ccb0eea8a706c4c34a16891f84e7b"); // employee password
@@ -110,7 +110,7 @@ public class CreateEmployee {
             pstmt.setString(9, "May");
             pstmt.setString(10, "1984");
             pstmt.setString(11, "sub.zero@gmail.com");
-            pstmt.setInt(12, 19000);
+            pstmt.setDouble(12, 19000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/subzero.png")));
             pstmt.setString(14, "normabrown");
             pstmt.setString(15, "827ccb0eea8a706c4c34a16891f84e7b");
@@ -127,7 +127,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Jul");
             pstmt.setString(10, "1969");
             pstmt.setString(11, "sonya.blade@outlook.com");
-            pstmt.setInt(12, 17000);
+            pstmt.setDouble(12, 17000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/sonya.png")));
             pstmt.setString(14, "sonyablade");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -144,7 +144,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Apr");
             pstmt.setString(10, "1971");
             pstmt.setString(11, "striker.watson@yahoo.co.uk");
-            pstmt.setInt(12, 24000);
+            pstmt.setDouble(12, 24000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/striker.png")));
             pstmt.setString(14, "strikerwatson");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -161,7 +161,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Apr");
             pstmt.setString(10, "1974");
             pstmt.setString(11, "smoke.martin@hotmail.com");
-            pstmt.setInt(12, 12000);
+            pstmt.setDouble(12, 12000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/smoke.png")));
             pstmt.setString(14, "smokemartin");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -178,7 +178,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Dec");
             pstmt.setString(10, "1973");
             pstmt.setString(11, "mileena.brooks@gmail.com");
-            pstmt.setInt(12, 23500);
+            pstmt.setDouble(12, 23500);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/mileena.png")));
             pstmt.setString(14, "mileenabrooks");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -195,7 +195,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Jun");
             pstmt.setString(10, "1993");
             pstmt.setString(11, "sindel.reed@gmail.com");
-            pstmt.setInt(12, 16000);
+            pstmt.setDouble(12, 16000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/sindel.png")));
             pstmt.setString(14, "sindelreed");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -212,7 +212,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Jun");
             pstmt.setString(10, "1971");
             pstmt.setString(11, "liu.kang@hotmail.com");
-            pstmt.setInt(12, 12000);
+            pstmt.setDouble(12, 12000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/liukang.png")));
             pstmt.setString(14, "liukang");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -229,7 +229,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Apr");
             pstmt.setString(10, "1979");
             pstmt.setString(11, "jade.hernandez@yahoo.co.uk");
-            pstmt.setInt(12, 12000);
+            pstmt.setDouble(12, 12000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/jade.png")));
             pstmt.setString(14, "jadehernandez");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -246,7 +246,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Jul");
             pstmt.setString(10, "1981");
             pstmt.setString(11, "kano.bailey@yahoo.co.uk");
-            pstmt.setInt(12, 12000);
+            pstmt.setDouble(12, 12000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/kano.png")));
             pstmt.setString(14, "kanobailey");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -263,7 +263,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Jul");
             pstmt.setString(10, "1987");
             pstmt.setString(11, "noob.saibot@outlook.com");
-            pstmt.setInt(12, 19000);
+            pstmt.setDouble(12, 19000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/noobsaibot.png")));
             pstmt.setString(14, "noobsaibot");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -280,7 +280,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Mar");
             pstmt.setString(10, "1984");
             pstmt.setString(11, "kobal.jones@gmail.com");
-            pstmt.setInt(12,14500);
+            pstmt.setDouble(12, 14500);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/kabal.png")));
             pstmt.setString(14, "kabaljones");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -297,7 +297,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Jan");
             pstmt.setString(10, "1989");
             pstmt.setString(11, "kitana.miller@hotmail.com");
-            pstmt.setInt(12,12000);
+            pstmt.setDouble(12, 12000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/kitana.png")));
             pstmt.setString(14, "kitanamiller");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -314,7 +314,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Dec");
             pstmt.setString(10, "1993");
             pstmt.setString(11, "reptile.evans@yahoo.co.uk");
-            pstmt.setInt(12, 22000);
+            pstmt.setDouble(12, 22000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/reptile.png")));
             pstmt.setString(14, "reptileevans");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -331,7 +331,7 @@ public class CreateEmployee {
             pstmt.setString(9, "May");
             pstmt.setString(10, "1992");
             pstmt.setString(11, "smoke.garcia@outlook.com");
-            pstmt.setInt(12, 12000);
+            pstmt.setDouble(12, 12000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/cyrax.png")));
             pstmt.setString(14, "smokegarcia");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -348,7 +348,7 @@ public class CreateEmployee {
             pstmt.setString(9, "May");
             pstmt.setString(10, "1979");
             pstmt.setString(11, "scorpion.perry@gmail.com");
-            pstmt.setInt(12, 22000);
+            pstmt.setDouble(12, 22000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/scorpion.png")));
             pstmt.setString(14, "scorpionperry");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -365,7 +365,7 @@ public class CreateEmployee {
             pstmt.setString(9, "May");
             pstmt.setString(10, "1981");
             pstmt.setString(11, "jax.jones@outlook.com");
-            pstmt.setInt(12, 16000);
+            pstmt.setDouble(12, 16000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/jax.png")));
             pstmt.setString(14, "jaxjames");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -382,7 +382,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Aug");
             pstmt.setString(10, "1985");
             pstmt.setString(11, "kung.lao@hotmail.com");
-            pstmt.setInt(12, 17500);
+            pstmt.setDouble(12, 17500);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/kunglao.png")));
             pstmt.setString(14, "kunglao");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -399,7 +399,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Sep");
             pstmt.setString(10, "1984");
             pstmt.setString(11, "nightwolf.lee@gmail.com");
-            pstmt.setInt(12, 12000);
+            pstmt.setDouble(12, 12000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/nightwolf.png")));
             pstmt.setString(14, "nightwolflee");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
@@ -416,7 +416,7 @@ public class CreateEmployee {
             pstmt.setString(9, "Oct");
             pstmt.setString(10, "1980");
             pstmt.setString(11, "shang.tsung@gmail.com");
-            pstmt.setInt(12, 12000);
+            pstmt.setDouble(12, 12000);
             pstmt.setBinaryStream(13, savePic2DB(new File("src/res/images/EmployeePictures/shangtsung.png")));
             pstmt.setString(14, "shangtsung");
             pstmt.setString(15, "81dc9bdb52d04dc20036dbd8313ed055");
